--- pysim/models/pingpong_oop/config.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-
-class Config(BaseModel):
-    interval: float
-    channel_delay: float
-    service_delay: float
-    loss_prob: float
-    max_pings: int | None = None
