+++ pysim/models/pingpong_oop/objects.py
@@ -1,6 +1,14 @@
 from pydantic import BaseModel, Field
 
 
+class Config(BaseModel):
+    interval: float
+    channel_delay: float
+    service_delay: float
+    loss_prob: float
+    max_pings: int | None = None
+
+
 class Result(BaseModel):
     avg_interval: float = Field(..., description="Average ping interval")
     avg_delay: float = Field(..., description="Average transmission delay")
